+++ tomcat/src/main/java/org/apache/coyote/http11/auth/AuthService.java
@@ -2,8 +2,8 @@ package org.apache.coyote.http11.auth;
 
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import org.apache.coyote.http11.request.body.RequestBody;
-import org.apache.coyote.http11.request.header.RequestHeader;
+import org.apache.coyote.http11.request.RequestBody;
+import org.apache.coyote.http11.request.RequestHeader;
 import org.apache.coyote.http11.request.line.Protocol;
 import org.apache.coyote.http11.request.line.RequestLine;
 import org.apache.coyote.http11.response.Location;
@@ -22,7 +22,7 @@ public class AuthService {
     private static final String LOGIN_PAGE = "/login.html";
     private static final String CONFLICT_PAGE = "/409.html";
     private static final String UNAUTHORIZED_PAGE = "/401.html";
-    private static final String COOKIE_NAME = "JSESSIONID";
+    public static final String SESSION_KEY = "JSESSIONID";
 
     private final SessionRepository sessionRepository = new SessionRepository();
 
@@ -30,7 +30,7 @@ public class AuthService {
         Protocol protocol = requestLine.protocol();
         if (requestLine.method().isGet()) {
             final Cookie cookie = requestHeader.getCookie();
-            final Session session = sessionRepository.getSession(cookie.get(COOKIE_NAME));
+            final Session session = sessionRepository.getSession(cookie.get(SESSION_KEY));
             if (session == null) {
                 return ResponseEntity.getCookieNullResponseEntity(protocol, OK, Location.from(LOGIN_PAGE));
             }
@@ -49,7 +49,7 @@ public class AuthService {
     private ResponseEntity getSuccessLoginResponse(final User user, final Protocol protocol) {
         ResponseEntity response = ResponseEntity.getCookieNullResponseEntity(protocol, FOUND,
                 Location.from(INDEX_PAGE));
-        String jsessionid = response.getHttpCookie().get("JSESSIONID");
+        String jsessionid = response.getHttpCookie().get(SESSION_KEY);
         final Session session = Session.from(jsessionid);
         session.setAttribute("user", user);
         sessionRepository.create(session);
